Move env_file_rust_log to logging so we don't have a feature flag combo that breaks
